Split the main function to main.dart
